Remove unnecessary functions, rearrange, generally cleanup
