+++ configure.ac
@@ -1,5 +1,5 @@
 #
-# semigroups: Methods for semigroups
+# Semigroups GAP package
 #
 # This file is part of the build system of a GAP kernel extension.
 # Requires GNU autoconf, GNU automake and GNU libtool.
@@ -10,7 +10,7 @@ dnl ## Setup autoconf
 dnl ##
 AC_PREREQ([2.68])
 AC_INIT([semigroups], m4_esyscmd([tr -d '\n' < VERSION]))
-AC_CONFIG_SRCDIR([src/semigroups.cc])
+AC_CONFIG_SRCDIR([src/gap.cc])
 AC_CONFIG_HEADER([src/pkgconfig.h:cnf/pkgconfig.h.in])
 AC_CONFIG_MACRO_DIR([m4])
 AC_CONFIG_AUX_DIR([cnf])
