--- src/boolean.h
@@ -1,19 +1,0 @@
-/*
- * Semigroups GAP package
- *
- * This file contains some methods for Boolean matrices
- *
- */
-
-#ifndef SEMIGROUPS_GAP_BOOLEAN_H
-#define SEMIGROUPS_GAP_BOOLEAN_H 1
-
-#include "src/compiled.h"
-
-#include "types.h"
-
-#include <assert.h>
-
-Obj IS_COL_TRIM_BOOLEAN_MAT (Obj, Obj);
-
-#endif
