+++ src/fropin.hh
@@ -1,0 +1,12 @@
+/*
+ * Semigroups GAP package
+ *
+ * This file contains... 
+ *
+ */
+
+#include "../../../src/compiled.h"
+
+Obj enumerate_semigroup (Obj self, Obj data, Obj limit, Obj lookfunc, Obj looking);
+Obj SCC_UNION_LEFT_RIGHT_CAYLEY_GRAPHS (Obj, Obj, Obj);
+Obj FIND_HCLASSES (Obj, Obj, Obj);
