+++ src/interface.cc
@@ -18,7 +18,7 @@
 #include "converter.h"
 #include "data.h"
 #include "interface.h"
-#include "types.h"
+#include "gap.hh"
 
 #include "semigroups++/semigroups.h"
 
@@ -129,7 +129,7 @@ Obj SEMIGROUP_ADD_GENERATORS (Obj self, Obj data, Obj coll_gap) {
   Converter* converter = data_converter(data);
   std::unordered_set<Element*> coll;
 
-  for(size_t i = 1; i <= (size_t) LEN_PLIST(coll_gap); i++) {
+  for (size_t i = 1; i <= (size_t) LEN_PLIST(coll_gap); i++) {
     coll.insert(converter->convert(ELM_PLIST(coll_gap, i),
                                    semigroup->degree())->really_copy());
   }
@@ -137,7 +137,7 @@ Obj SEMIGROUP_ADD_GENERATORS (Obj self, Obj data, Obj coll_gap) {
 
   Obj gens = ElmPRec(data, RNam_gens); // TODO make this safe
 
-  for(size_t i = 0; i < semigroup->nrgens(); i++) {
+  for (size_t i = 0; i < semigroup->nrgens(); i++) {
     AssPlist(gens, i + 1, converter->unconvert(semigroup->gens()->at(i)));
   }
