--- src/types.cc
@@ -1,39 +1,0 @@
-/*
- * Semigroups GAP package
- *
- * This file contains types of semigroups for use in the Semigroups kernel
- * module.
- *
- */
-
-#include "types.h"
-
-Obj HTValue;
-Obj HTAdd;
-Obj infinity;
-Obj Ninfinity;
-Obj IsBipartition;
-Obj BipartTypes;
-Obj BipartitionType;
-Obj BlocksType;
-Obj IsBooleanMat;
-Obj BooleanMatType;
-Obj IsMatrixOverSemiring;
-Obj IsMaxPlusMatrix;
-Obj MaxPlusMatrixType;
-Obj IsMinPlusMatrix;
-Obj MinPlusMatrixType;
-Obj IsTropicalMatrix;
-Obj IsTropicalMinPlusMatrix;
-Obj TropicalMinPlusMatrixType;
-Obj IsTropicalMaxPlusMatrix;
-Obj TropicalMaxPlusMatrixType;
-Obj IsProjectiveMaxPlusMatrix;
-Obj ProjectiveMaxPlusMatrixType;
-Obj IsNTPMatrix;
-Obj NTPMatrixType;
-Obj IsIntegerMatrix;
-Obj IntegerMatrixType;
-Obj IsPBR;
-Obj PBRTypes;
-Obj PBRType;
