+++ tst/standard/blocks.tst
@@ -156,12 +156,12 @@ gap> x := Bipartition([[1, 4, 7, 8, 12, -1, -5, -14],
 > [-10, -13, -15]]);;
 gap> blocks := LeftBlocks(x);;
 gap> CanonicalBlocks(blocks);
-<blocks: [ 1*, 2* ], [ 3*, 6*, 5*, 4* ], [ 7*, 10*, 9*, 8* ], [ 11*, 15*, 14*,\
- 13*, 12* ], [ 16*, 20*, 19*, 18*, 17* ]>
+<blocks: [ 1*, 2* ], [ 3*, 4*, 5*, 6* ], [ 7*, 8*, 9*, 10* ], [ 11*, 12*, 13*,\
+ 14*, 15* ], [ 16*, 17*, 18*, 19*, 20* ]>
 gap> blocks := RightBlocks(x);;
 gap> CanonicalBlocks(blocks);
-<blocks: [ 1*, 2* ], [ 3*, 4* ], [ 5*, 7*, 6* ], [ 8*, 10*, 9* ], [ 11*, 13*, \
-12* ], [ 14, 16, 15 ], [ 17, 20, 19, 18 ]>
+<blocks: [ 1*, 2* ], [ 3*, 4* ], [ 5*, 6*, 7* ], [ 8*, 9*, 10* ], [ 11*, 12*, \
+13* ], [ 14, 15, 16 ], [ 17, 18, 19, 20 ]>
 
 #T# SEMIGROUPS_UnbindVariables
 gap> Unbind(S);
